Refactor EventForm component to use DatePicker component
This commit refactors the EventForm component to use the DatePicker component for date and time input fields. The changes include updating the JSX syntax and adding the necessary props for the DatePicker component. The commit also removes the redundant code for displaying the (UTC) label when the component is not hydrated.